Slightly refactor storage code in preparation for pgsql
